requirements.txt
asgiref==3.2.7
boto3==1.12.31
botocore==1.15.31
dj-database-url==0.5.0
Django==3.0.4
django-resized==0.3.9
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
sendgrid==5.3.0
